add data subcommands for gc, sync, compact. also alias putData and getData
